2025/BLOCKCHAIN/Vastavikamaina-Token/admin/src/python-service-vastavikamaina-token/requirements.txt
web3==6.18.0
python-dotenv==1.0.1
psutil